Funcionalidad para cambiar la imagen a un libro de la biblioteca y guardarlo en almacenamiento Firebase.

+++ app/src/main/java/com/example/practica2/Actividades/MainActivity.java
@@ -322,12 +322,14 @@ public class MainActivity extends AppCompatActivity implements DialogoImagen.Lis
                                 Toast.makeText(MainActivity.this, R.string.errorEscaneo, Toast.LENGTH_SHORT).show();
                             }
                             boolean deleted = false;
-                            try {
-                                deleted = fichImg.delete();
-                            } catch (SecurityException e) {
-                            }
-                            if (!deleted) {
-                                fichImg.deleteOnExit();
+                            if(fichImg!=null) {
+                                try {
+                                    deleted = fichImg.delete();
+                                } catch (SecurityException e) {
+                                }
+                                if (!deleted) {
+                                    fichImg.deleteOnExit();
+                                }
                             }
                         }
                     })
